@@ -36,23 +36,24 @@ Gem::Specification.new do |spec|
   spec.add_dependency 'faraday_middleware', '~> 0.11'
   spec.add_dependency 'dotenv', '~> 2.2'
   spec.add_dependency 'climate_control', '~> 0'
-  spec.add_dependency 'require_all', '~> 1.3'
-  spec.add_dependency 'bibtex-ruby', '~> 4.0'
+  spec.add_dependency 'require_all', '~> 2.0'
+  spec.add_dependency 'bibtex-ruby', '~> 5.0', '>= 5.0.0'
   spec.add_dependency 'citeproc', '>= 1.0.4', '< 2.0'
   spec.add_dependency 'csl', '~> 1.4'
   spec.add_dependency 'citeproc-ruby', '~> 1.0', '>= 1.0.2'
   spec.add_dependency 'csl-styles', '~> 1.0', '>= 1.0.1.5'
-  spec.add_dependency 'activesupport', '~> 5.0'
-  spec.add_dependency 'net-http-persistent', '~> 2.9'
-  spec.add_dependency 'sanitize', '~> 4.5', '>= 4.5.0'
+  spec.add_dependency 'activesupport', '~> 5.2'
+  spec.add_dependency 'net-http-persistent', '~> 3.1'
+  spec.add_dependency 'sanitize', '~> 5.0'
+  spec.add_dependency 'public_suffix', '~>4.0'
 
   spec.add_development_dependency 'bundler', '~> 1.13'
-  spec.add_development_dependency 'rake', '~> 12.0'
+  spec.add_development_dependency 'rake', '~> 12.3'
   spec.add_development_dependency 'minitest', '~> 5.0'
-  spec.add_development_dependency 'simplecov', '~> 0'
+  spec.add_development_dependency 'simplecov', '~> 0.17.0'
   spec.add_development_dependency 'codecov', '~> 0.1'
-  spec.add_development_dependency 'vcr', '~> 2.8'
-  spec.add_development_dependency 'minitest-vcr', '~> 0.1.1'
-  spec.add_development_dependency 'webmock', '~> 1.17'
+  spec.add_development_dependency 'vcr', '~> 5.0', '>= 5.0.0'
+  spec.add_development_dependency 'minitest-vcr', '~> 1.4', '>= 1.4.0'
+  spec.add_development_dependency 'webmock', '~> 3.6'
 
 end
