@@ -1,5 +1,5 @@
 module EBSCO
   module EDS
-    VERSION = '1.0.7'
+    VERSION = '1.0.8'
   end
 end
