@@ -3,7 +3,7 @@ require_relative 'test_helper'
 class EdsApiTests < Minitest::Test
 
   def test_next_page
-    VCR.use_cassette('paging_test/profile_1/test_next_page') do
+    VCR.use_cassette('paging_test/profile_1/test_next_page', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({query: 'economic development'})
       assert results.page_number == 1
@@ -14,7 +14,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_get_page
-    VCR.use_cassette('paging_test/profile_1/test_get_page') do
+    VCR.use_cassette('paging_test/profile_1/test_get_page', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({query: 'economic development'})
       assert results.page_number == 1
@@ -25,7 +25,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_prev_page
-    VCR.use_cassette('paging_test/profile_1/test_prev_page') do
+    VCR.use_cassette('paging_test/profile_1/test_prev_page', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({query: 'economic development'})
       assert results.page_number == 1
@@ -38,7 +38,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_prev_page_before_one
-    VCR.use_cassette('paging_test/profile_1/test_prev_page_before_one') do
+    VCR.use_cassette('paging_test/profile_1/test_prev_page_before_one', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({query: 'economic development'})
       assert results.page_number == 1
@@ -49,7 +49,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_next_page_past_last_page
-    VCR.use_cassette('paging_test/profile_1/test_next_page_past_last_page') do
+    VCR.use_cassette('paging_test/profile_1/test_next_page_past_last_page', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({query: 'economic development'})
       assert results.page_number == 1
@@ -61,7 +61,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_next_page_with_only_one_page_of_results
-    VCR.use_cassette('paging_test/profile_1/test_next_page_with_only_one_page_of_results') do
+    VCR.use_cassette('paging_test/profile_1/test_next_page_with_only_one_page_of_results', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({query: 'megaenzymes', results_per_page: 100})
       assert results.page_number == 1
@@ -72,7 +72,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_move_page
-    VCR.use_cassette('paging_test/profile_1/test_move_page') do
+    VCR.use_cassette('paging_test/profile_1/test_move_page', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({query: 'economic development'})
       assert results.page_number == 1
@@ -83,7 +83,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_reset_page
-    VCR.use_cassette('paging_test/profile_1/test_reset_page') do
+    VCR.use_cassette('paging_test/profile_1/test_reset_page', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({query: 'economic development'})
       assert results.page_number == 1
@@ -96,7 +96,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_solr_beyond_250_results
-    VCR.use_cassette('paging_test/profile_1/test_beyond_250_results') do
+    VCR.use_cassette('paging_test/profile_1/test_beyond_250_results', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({'q' => 'cats', 'page' => 5, 'per_page' => 100, 'search_field' => 'all_fields'})
       assert results.page_number == 5
@@ -105,7 +105,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_solr_beyond_250_results_with_limiter
-    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_limiter') do
+    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_limiter', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({'q' => 'cats',
         'f' => {'eds_search_limiters_facet'=>['Peer Reviewed']},
@@ -116,7 +116,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_solr_beyond_250_results_with_source_type_facet_only
-    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_source_type_facet_only') do
+    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_source_type_facet_only', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({'q' => 'cats',
                                 'f' => {'eds_publication_type_facet'=>['Books']},
@@ -127,7 +127,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_solr_beyond_250_results_with_content_provider_facet_only
-    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_content_provider_facet_only') do
+    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_content_provider_facet_only', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({'q' => 'cats',
                                 'f' => {'eds_content_provider_facet'=>['Academic Search Ultimate']},
@@ -138,7 +138,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_solr_beyond_250_results_with_source_type_facet_and_limiter
-    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_source_type_facet_and_limiter') do
+    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_source_type_facet_and_limiter', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({'q' => 'cats',
                                 'f' => {'eds_publication_type_facet'=>['Books'],
@@ -150,7 +150,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_solr_beyond_250_results_with_content_provider_facet_and_limiter
-    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_content_provider_facet_and_limiter') do
+    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_content_provider_facet_and_limiter', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({'q' => 'cats',
                                 'f' => {'eds_content_provider_facet'=>['Academic Search Ultimate'],
@@ -162,7 +162,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_solr_beyond_250_results_with_source_type_facet_and_language_facet
-    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_source_type_facet_and_language_facet') do
+    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_source_type_facet_and_language_facet', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({'q' => 'cats',
                               'f' => {'eds_language_facet' => ['english'],
@@ -174,7 +174,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_solr_beyond_250_results_with_content_provider_facet_and_language_facet
-    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_content_provider_facet_and_language_facet') do
+    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_content_provider_facet_and_language_facet', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({'q' => 'cats',
                             'f' => {'eds_content_provider_facet'=>['Academic Search Ultimate'],
@@ -186,7 +186,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_solr_beyond_250_results_with_content_provider_facet_and_source_type_facet
-    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_content_provider_facet_and_source_type_facet') do
+    VCR.use_cassette('paging_test/profile_1/test_solr_beyond_250_results_with_content_provider_facet_and_source_type_facet', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'eds-api'})
       results = session.search({'q' => 'cats',
                       'f' => {'eds_content_provider_facet'=>['Academic Search Ultimate'],
@@ -198,7 +198,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_solr_beyond_250_results_with_multiple_facets
-    VCR.use_cassette('paging_test/profile_2/test_solr_beyond_250_results_with_multiple_facets') do
+    VCR.use_cassette('paging_test/profile_2/test_solr_beyond_250_results_with_multiple_facets', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'edsapi'})
       results = session.search({'q' => 'cats',
                                 'f' => {'eds_language_facet'=>['english'],
@@ -211,7 +211,7 @@ class EdsApiTests < Minitest::Test
   end
 
   def test_solr_beyond_250_results_with_multiple_facets_and_limiter
-    VCR.use_cassette('paging_test/profile_2/test_solr_beyond_250_results_with_multiple_facets_and_limiter') do
+    VCR.use_cassette('paging_test/profile_2/test_solr_beyond_250_results_with_multiple_facets_and_limiter', :allow_playback_repeats => true) do
       session = EBSCO::EDS::Session.new({use_cache: false, profile: 'edsapi'})
       results = session.search({'q' => 'cats',
                                 'f' => {'eds_language_facet'=>['english'],
